requirements.txt
conan>=2.18
PyYAML
